src/functions/cards_processor/requirements.txt
google-cloud-storage==2.14.0
functions-framework==3.5.0
structlog==24.1.0
requests==2.31.0 